added compromises + bug fixes
now if an oppurtunity is found but can't be exploited because 1. the maximum transaction size is too small or 2. the maximum possible profit is too small, then it identifies which conversion limits the transaction size the most and updates the conversion rates so that that conversion now includes the 1st and second best conversion rate (thereby increasing the maximum transaction size through that conversion). It then checks if the new conversion rates present an oppurtunity and explores that oppurtunity. It will make up to 5 compromises before giving up on the conversion. Also bug fixes.