started experimenting with actual purchases
have now got the exchanges i want to operate on down, compromises and arbitrage oppurtunity detection are solid

the only problem is the time between when the conversion rates are loaded and when the purchases are actually made could mean we lose money at times, or leave small amounts of money in unstable currencies. This may or not be a huge problem, only way to know is by testing it out.

Also changed minimum transaction size hopefully there are still exchanges that will present oppurtunities. And also each exchange should have something like 20 dollars on it so that my money isnt stuck (I have money stuck on aax and btcalpha because the balance is less than 5 dollars so i cant transfer it anywhere without adding more money because of minimum transaction size.